Finished the task.
Took me a little longer than two hours, took me about double that. Kept referring to language reference to clean up the text, that was honestly the hardest part because getting the data then putting it in a table was fairly easy.

I'll add readme to suggest further improvements that I can make in the future.